Add creating url and change price to type BigDecimal

+++ src/main/java/com/bmxApp/repository/BasketProductRepository.java
@@ -1,5 +1,6 @@
 package com.bmxApp.repository;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -18,7 +19,7 @@ public interface BasketProductRepository extends JpaRepository<BasketProduct, In
 	List<BasketProduct> findByShopName(String shopName);
 
 	@Query(value = "Select (bt.quantity * pt.price) from BasketProduct bt inner join Product pt on bt.product = pt.id where pt.id=?1")
-	public float getTotalPriceForBasketProduct(int productId);
+	public BigDecimal getTotalPriceForBasketProduct(int productId);
 
 	@Query(value = "Select SUM(bt.quantity * pt.price) from BasketProduct bt inner join Product pt on bt.product = pt.id")
 	public float getTotalPrice();
